requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-crispy-forms==1.14.0
django-utils-six==2.0
djangorestframework==3.13.1
Pillow==9.0.1
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
